// src/api/types.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export type RequestMethod = "GET" | "POST" | "PATCH" | "DELETE" | "PUT";

// Generated by https://quicktype.io

export interface MondayUser {
  id: string
  name: string
  email: string
  is_guest: boolean
  created_at: string
}

export interface ItemBoard {
  id: string
  name: string
}

export interface IItem {
  id: string;
  name: string;
  state: string;
  created_at: string;
  column_values: ColumnValue[];
  updates: Update[];
  workspace: string;
  board: ItemBoard;
  group: Group;
  creator: {
    id: string;
    name: string;
  };
  workspace_id?: number | null;
}

export interface ColumnValue {
  id: string;
  title: string;
  value: null | string;
  type: string;
  text: string;
}

export interface Update {
  body: string;
  creator: Creator;
  created_at: string;
}

export interface Creator {
  name: string;
}

export interface ItemsPage {
  cursor: string | null,
  items: IItem[]
}

export interface IBoard {
  id: string;
  name: string;
  workspace_id: number | null;
  groups: Group[];
  items_page: ItemsPage;
}

export interface Group {
  id: string;
  title: string;
}

// Generated by https://quicktype.io

export interface IWorkspaces {
  id: number;
  name: string;
}

// Generated by https://quicktype.io

export interface BoardColumns {
  boards: BoardColumn[];
}

export interface BoardColumn {
  columns: Column[];
}

export interface Column {
  id: string;
  title: string;
  settings_str: string;
  description: null;
  type: string;
}